Feat: Rename the files in the jsonjoy-builder directory to lowercase. #10866
